database-knot-agent-main/requirements.txt
mysql-connector-python==8.0.29
paho-mqtt